Creating PentagonGridCpp
